+++ src/config/theme/properties.rs
@@ -110,6 +110,18 @@ pub enum StatusPropertyFile {
     Duration,
     Crossfade,
     Bitrate,
+    QueueLength {
+        #[serde(default = "defaults::default_thousands_separator")]
+        thousands_separator: String,
+    },
+    QueueTimeTotal {
+        #[serde(default = "defaults::default_time_unit_separator")]
+        separator: String,
+    },
+    QueueTimeRemaining {
+        #[serde(default = "defaults::default_time_unit_separator")]
+        separator: String,
+    },
 }
 
 #[derive(Debug, Clone, Display, Hash, Eq, PartialEq)]
@@ -155,6 +167,15 @@ pub enum StatusProperty {
     Duration,
     Crossfade,
     Bitrate,
+    QueueLength {
+        thousands_separator: String,
+    },
+    QueueTimeTotal {
+        separator: String,
+    },
+    QueueTimeRemaining {
+        separator: String,
+    },
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -416,6 +437,15 @@ impl TryFrom<StatusPropertyFile> for StatusProperty {
                     .map(|s| -> Result<_> { s.to_config_or(None, None) })
                     .transpose()?,
             },
+            StatusPropertyFile::QueueLength { thousands_separator } => {
+                StatusProperty::QueueLength { thousands_separator }
+            }
+            StatusPropertyFile::QueueTimeTotal { separator } => {
+                StatusProperty::QueueTimeTotal { separator }
+            }
+            StatusPropertyFile::QueueTimeRemaining { separator } => {
+                StatusProperty::QueueTimeRemaining { separator }
+            }
         })
     }
 }